+++ components/Layouts/HeaderComponent.jsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import React from 'react'
-import styles from '../styles/Header.module.css'
+import styles from '../../styles/layouts/Header.module.css'
 
 
 export default function HeaderComponent({ title, titleLink, callToAction, callToActionLink }) {
