--- pages/blog.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function blog() {
-  return (
-    <div>Blog main page in development</div>
-  )
-}
