+++ pages/es.js
@@ -1,12 +1,8 @@
 import Head from "next/head";
-import styles from "../styles/Home.module.css";
+import styles from "../styles/pages/Home.module.css";
 import HeaderComponent from "../components/HeaderComponent";
 import FooterComponent from "../components/FooterComponent";
 import BannerComponent from "../components/BannerComponent";
-import AboutEsComponent from "../components/AboutEsComponent";
-import GaleryComponent from '../components/GaleryComponent';
-import PhoneGalery from '../components/PhoneGalery';
-import TitleComponent from '../components/TitleComponent';
 import { BsArrowRightShort } from 'react-icons/bs';
 import Link from "next/link";
 import Image from "next/image";
