--- styles/About.module.css
@@ -1,18 +1,0 @@
-.about {
-    text-align: left;
-    max-width: 700px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.title {
-    font-size: 40px;
-    color: #226E93;
-    margin: 30px 0;
-}
-
-.text {
-    line-height: 35px;
-    font-size: 18px;
-    color: #505050;
-}
