--- styles/Box.module.css
@@ -1,32 +1,0 @@
-.box {
-    display: flex;
-    align-items: flex-end;
-    justify-content: flex-end;
-}
-
-.text {
-    padding: 10px 15px 10px 15px;
-    color: white;
-    font-size: 20px;
-    background-color: rgba(159, 159, 159, 0.637);
-    transition: 0.3s;
-    position: absolute;
-    border-bottom-right-radius: 10px;
-    border-top-left-radius: 10px;
-    max-width: 280px;
-}
-
-.box:hover {
-    cursor: pointer;
-}
-/* 
-.box:hover .text {
-    font-size: 25px;
-} */
-
-.box:hover .image {
-    height: 95%;
-    width: 95%;
-    /* -webkit-filter: blur(0.7px);
-    filter: blur(0.7px); */
-}
