--- styles/Galery.module.css
@@ -1,5 +1,0 @@
-.galery div {
-    display: flex;
-    flex-direction: row;
-    margin: 3px;
-}
