--- styles/GaleryPhone.module.css
@@ -1,4 +1,0 @@
-.galery div {
-    margin-top: 10px;
-    margin-bottom: 10px;
-}
