--- styles/Portfolio.module.css
@@ -1,89 +1,0 @@
-.line {
-    border-bottom: 1px #b4b4b4 solid;
-    margin-right: 10px;
-    margin-left: 10px;
-}
-
-.body {
-    max-width: 960px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.filters {
-    margin: 37px 10px 30px 10px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.title {
-    font-size: 20px;
-    color: #226E93;
-    cursor: pointer;
-}
-
-.searchbar {
-    display: flex;
-    flex-direction: row;
-}
-
-.searchinput {
-    padding: 10px 10px 9px 10px;
-    /* width: 400px; */
-    border-top: 1px solid #9c9c9c;
-    border-right: none;
-    border-bottom: 1px solid #9c9c9c;
-    border-left: 1px solid #9c9c9c;
-    border-top-left-radius: 5px;
-    border-bottom-left-radius: 5px;
-}
-
-.searchbutton {
-    padding: 10px 10px 8px 10px;
-    background-color: white;
-    border-top: 1px solid #9c9c9c;
-    border-right: 1px solid #9c9c9c;
-    border-bottom: 1px solid #9c9c9c;
-    border-left: none;
-    border-top-right-radius: 5px;
-    border-bottom-right-radius: 5px;
-    padding-right: 10px;
-    cursor: pointer;
-}
-
-.list {
-    margin-bottom: 50px;
-    text-align: left;
-}
-
-.element {
-    display: inline-block;
-    margin: 10px;
-}
-
-@media (max-width: 982px) {
-    .body {
-        max-width: 640px;
-    }
-}
-
-@media (max-width: 657px) {
-    .body {
-        max-width: 320px;
-    }
-
-    .filters {
-        flex-direction: column;
-        margin-bottom: 15px;
-    }
-
-    .searchinput {
-        width: 265px;
-    }
-
-    .title {
-        margin-bottom: 30px;
-    }
-}
