--- styles/Title.module.css
@@ -1,5 +1,0 @@
-.title {
-    font-size: 20px;
-    color: #226E93;
-    margin: 30px auto;
-}
